pyproject.toml
[project]
name = "openclaw-telegram-bot"
version = "0.3.0"
description = "AI-powered Telegram bot with multi-provider LLM support"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "groq>=0.4.0",
    "ollama>=0.1.0",
    "python-telegram-bot>=21.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "flask>=3.0.0",
    "httpx>=0.27.0",
    "feedparser>=6.0.0",
    "yt-dlp>=2024.0.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.100.0",
]

[project.scripts]
openclaw = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
